+++ assets/test/defaultPackage/Tests/PackageTests2/PackageTests2.swift
@@ -1,0 +1,13 @@
+#if swift(>=6.0)
+import Testing
+
+@Test func secondTargetTestPassing() {
+  #expect(1 == 1)
+}
+
+@Suite
+struct SecondTargetSuite {
+  @Test
+  func testPassing() throws {}
+}
+#endif